feat: add override for button color
