--- src/components/icons/lwcIconColored.ts
@@ -1,13 +1,0 @@
-import { svg } from "lit";
-
-export const lwcIconColored = svg`<svg width="33" height="32" viewBox="0 0 33 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M32.3023 17.4161C31.0174 18.6948 29.2424 18.3883 27.9635 17.1038C27.9635 17.1038 27.9635 17.1038 27.9633 17.1036L17.6133 6.70709L22.4634 1.83601C23.8679 0.425247 26.251 0.862027 27.0638 2.67931L32.4596 16.7343C32.5483 16.9652 32.4924 17.2268 32.317 17.4015L28.9089 20.7935L32.3021 17.4161H32.3023Z" fill="#472459"/>
-<path d="M27.9623 17.1035C27.9623 17.1035 27.9623 17.1035 27.9621 17.1033L14.4236 3.5038C13.1445 2.2191 11.0662 2.21428 9.78116 3.49301L1.46688 11.7681C0.18183 13.047 0.177204 15.1251 1.45609 16.41L14.9944 30.0094C16.2735 31.2941 18.3518 31.2989 19.6369 30.0202L22.3103 27.3594L23.0936 26.5799L20.5946 24.0718C18.7462 25.2699 16.2498 25.0568 14.6341 23.4311L12.9336 21.72C12.6919 21.4767 12.693 21.0876 12.9363 20.8459L13.7732 20.0143L11.6715 17.8996C11.3421 17.5681 11.2904 17.0349 11.5875 16.676C11.9273 16.2639 12.5404 16.2446 12.9097 16.6163L15.038 18.7578L16.7065 17.1001L14.6048 14.9854C14.2754 14.6538 14.2237 14.1207 14.5235 13.7593C14.8658 13.3499 15.4764 13.3279 15.8457 13.6995L17.974 15.841L18.8135 15.0068C19.0568 14.7651 19.446 14.7662 19.6878 15.0095L21.3883 16.7205C23.0041 18.3464 23.2013 20.8441 21.9909 22.6847L24.4887 25.1915L25.9121 23.7747L27.9511 21.7452L32.301 17.416C31.0161 18.6947 29.241 18.3882 27.9619 17.1037L27.9623 17.1035Z" fill="url(#paint0_linear_433_409)"/>
-<defs>
-<linearGradient id="paint0_linear_433_409" x1="16.4005" y1="2.53711" x2="16.4005" y2="30.9761" gradientUnits="userSpaceOnUse">
-<stop stop-color="#FFCA4A"/>
-<stop offset="1" stop-color="#F7931A"/>
-</linearGradient>
-</defs>
-</svg>
-`
