+++ website/src/components/Header.tsx
@@ -1,11 +1,8 @@
 import '@getalby/lightning-wallet-connect';
-import {useAppStore} from '../AppStore';
 import {BCLogo} from '../icons/BCLogo';
-import {SunIcon, MoonIcon} from '@heroicons/react/24/solid';
+import {DarkModeToggle} from './DarkModeToggle';
 
 export function Header() {
-  const store = useAppStore();
-
   return (
     <div className="flex w-full justify-between items-center max-sm:flex-col max-sm:gap-4">
       <div className="flex gap-4 justify-center items-center">
@@ -15,16 +12,7 @@ export function Header() {
       <div className="flex justify-end items-center gap-4">
         {/* @ts-ignore */}
         <lwc-button />
-        <label className="swap swap-rotate hover:brightness-110 active:scale-95">
-          <input
-            type="checkbox"
-            checked={store.isDarkMode}
-            onChange={(event) => store.setDarkMode(event.target.checked)}
-          />
-
-          <SunIcon className="swap-on w-8 h-8" />
-          <MoonIcon className="swap-off w-8 h-8" />
-        </label>
+        <DarkModeToggle />
       </div>
     </div>
   );
